airBNB
airBnB

For the main page which is airBnB.html we can see the login button, and when we press it, the page moves login page.
We can sign-up and log-in at the same page.
When we sign-up, the information will be stored in localstorage, and we can log-in with the same information that is already stored, and with log-in the page moves to the main page, and we can see the Log-in button has been changed to Log-out.
Even at the bottom of the main page, we can see the difference between before and after login.

When we press  Plan & book button, it goes search page and the datepicker, we can see its working. The initial date cannot be bigger than final date.
After press the search button it moves to the bookingpage-rooms.html.
In the search bar, we can search the place that we may want to go, and since this web page has only for place that for search, which are toronto, quebec, montreal, and vancouver.
We can input one of the places name and see the result.
If you choose one of the rooms, it moves to the detail page of the room that has been chosen.
And in that page we can book the room and there’s also datepicker’s working we can see, and after booking, there will be an alert that a notification of the booking.
if we loged-in, we can see the email address in the notification